refs #6: Fix a memory allocation bug in GPU IO buffers.
 * We need to return the pointer "before" adding the newly allocated
   size (in lib/mempool.hh).

 * Now GPU does not crash, but it still hangs after processing the first
   offload task.

+++ engines/cuda/mempool.hh
@@ -36,7 +36,7 @@ public:
         size_t offset;
         int ret = _alloc(size, &offset);
         if (ret == 0)
-            return (void *) ((uint8_t *) base_ + offset);
+            return (void *) ((uint8_t *) base_ + (uintptr_t) offset);
         return NULL;
     }
 
@@ -80,7 +80,7 @@ public:
         size_t offset;
         int ret = _alloc(size, &offset);
         if (ret == 0)
-            return (void *) ((uint8_t *) base_ + offset);
+            return (void *) ((uint8_t *) base_ + (uintptr_t) offset);
         return NULL;
     }
